move button and notifications to inside the new component
