+++ dotcom-rendering/src/components/FollowButtons.stories.tsx
@@ -1,11 +1,6 @@
-import { css } from '@emotion/react';
 import { splitTheme } from '../../.storybook/decorators/splitThemeDecorator';
 import { ArticleDesign, ArticleDisplay, Pillar } from '../lib/articleFormat';
-import {
-	FollowNotificationsButton,
-	FollowTagButton,
-	FollowTagButtonVariant,
-} from './FollowButtons';
+import { FollowNotificationsButton, FollowTagButton } from './FollowButtons';
 
 export default {
 	component: [FollowNotificationsButton, FollowTagButton],
@@ -75,79 +70,3 @@ export const FollowContributorBothStates = () => {
 	);
 };
 FollowContributorBothStates.decorators = [splitTheme()];
-
-const variantContainerStyles = css`
-	display: flex;
-	gap: 16px;
-	margin-bottom: 16px;
-`;
-
-export const VariantFollowTagButtonBothStates = () => {
-	return (
-		<div css={variantContainerStyles}>
-			<FollowTagButtonVariant
-				isFollowing={false}
-				onClickHandler={() => undefined}
-			/>
-			<FollowTagButtonVariant
-				isFollowing={true}
-				onClickHandler={() => undefined}
-			/>
-		</div>
-	);
-};
-VariantFollowTagButtonBothStates.storyName = 'Variant Button - Both States';
-VariantFollowTagButtonBothStates.decorators = [
-	splitTheme([
-		{
-			display: ArticleDisplay.Standard,
-			design: ArticleDesign.Standard,
-			theme: Pillar.News,
-		},
-	]),
-];
-
-export const VariantFollowTagButtonAllPillars = () => {
-	return (
-		<div css={variantContainerStyles}>
-			<FollowTagButtonVariant
-				isFollowing={false}
-				onClickHandler={() => undefined}
-			/>
-			<FollowTagButtonVariant
-				isFollowing={true}
-				onClickHandler={() => undefined}
-			/>
-		</div>
-	);
-};
-VariantFollowTagButtonAllPillars.storyName = 'Variant Button - All Pillars';
-VariantFollowTagButtonAllPillars.decorators = [
-	splitTheme([
-		{
-			display: ArticleDisplay.Standard,
-			design: ArticleDesign.Standard,
-			theme: Pillar.News,
-		},
-		{
-			display: ArticleDisplay.Standard,
-			design: ArticleDesign.Standard,
-			theme: Pillar.Opinion,
-		},
-		{
-			display: ArticleDisplay.Standard,
-			design: ArticleDesign.Standard,
-			theme: Pillar.Sport,
-		},
-		{
-			display: ArticleDisplay.Standard,
-			design: ArticleDesign.Standard,
-			theme: Pillar.Culture,
-		},
-		{
-			display: ArticleDisplay.Standard,
-			design: ArticleDesign.Standard,
-			theme: Pillar.Lifestyle,
-		},
-	]),
-];
